requirements.txt
kafka-python==2.0.2
pysolr==3.9.0
